+++ src/main/kotlin/dev/aaronhowser/mods/geneticsresequenced/registry/ModBlocks.kt
@@ -11,6 +11,7 @@ import dev.aaronhowser.mods.geneticsresequenced.block.machine.dna_extractor.DnaE
 import dev.aaronhowser.mods.geneticsresequenced.block.machine.incubator.IncubatorBlock
 import dev.aaronhowser.mods.geneticsresequenced.block.machine.incubator_advanced.AdvancedIncubatorBlock
 import dev.aaronhowser.mods.geneticsresequenced.block.machine.plasmid_infuser.PlasmidInfuserBlock
+import dev.aaronhowser.mods.geneticsresequenced.block.machine.plasmid_injector.PlasmidInjectorBlock
 import net.minecraft.world.level.block.Block
 import net.minecraft.world.level.block.state.BlockBehaviour
 import net.neoforged.neoforge.registries.DeferredBlock
@@ -23,17 +24,27 @@ object ModBlocks {
 
     val BIOLUMINESCENCE_BLOCK: DeferredBlock<BioluminescenceBlock> =
         registerBlock("bioluminescence") { BioluminescenceBlock() }
-    val ANTI_FIELD_BLOCK: DeferredBlock<AntiFieldBlock> = registerBlock("anti_field_block") { AntiFieldBlock() }
+    val ANTI_FIELD_BLOCK: DeferredBlock<AntiFieldBlock> =
+        registerBlock("anti_field_block") { AntiFieldBlock() }
 
-    val COAL_GENERATOR: DeferredBlock<CoalGeneratorBlock> = registerBlock("coal_generator") { CoalGeneratorBlock() }
-    val CELL_ANALYZER: DeferredBlock<CellAnalyzerBlock> = registerBlock("cell_analyzer") { CellAnalyzerBlock() }
-    val DNA_EXTRACTOR: DeferredBlock<Block> = registerBlock("dna_extractor") { DnaExtractorBlock() }
-    val DNA_DECRYPTOR: DeferredBlock<DnaDecryptorBlock> = registerBlock("dna_decryptor") { DnaDecryptorBlock() }
-    val BLOOD_PURIFIER: DeferredBlock<BloodPurifierBlock> = registerBlock("blood_purifier") { BloodPurifierBlock() }
-    val PLASMID_INFUSER: DeferredBlock<PlasmidInfuserBlock> = registerBlock("plasmid_infuser") { PlasmidInfuserBlock() }
-    val PLASMID_INJECTOR: DeferredBlock<Block> = registerBlock("plasmid_injector")
-    val INCUBATOR: DeferredBlock<IncubatorBlock> = registerBlock("incubator") { IncubatorBlock() }
-    val ADVANCED_INCUBATOR: DeferredBlock<Block> = registerBlock("advanced_incubator") { AdvancedIncubatorBlock() }
+    val COAL_GENERATOR: DeferredBlock<CoalGeneratorBlock> =
+        registerBlock("coal_generator") { CoalGeneratorBlock() }
+    val CELL_ANALYZER: DeferredBlock<CellAnalyzerBlock> =
+        registerBlock("cell_analyzer") { CellAnalyzerBlock() }
+    val DNA_EXTRACTOR: DeferredBlock<DnaExtractorBlock> =
+        registerBlock("dna_extractor") { DnaExtractorBlock() }
+    val DNA_DECRYPTOR: DeferredBlock<DnaDecryptorBlock> =
+        registerBlock("dna_decryptor") { DnaDecryptorBlock() }
+    val BLOOD_PURIFIER: DeferredBlock<BloodPurifierBlock> =
+        registerBlock("blood_purifier") { BloodPurifierBlock() }
+    val PLASMID_INFUSER: DeferredBlock<PlasmidInfuserBlock> =
+        registerBlock("plasmid_infuser") { PlasmidInfuserBlock() }
+    val PLASMID_INJECTOR: DeferredBlock<PlasmidInjectorBlock> =
+        registerBlock("plasmid_injector") { PlasmidInjectorBlock() }
+    val INCUBATOR: DeferredBlock<IncubatorBlock> =
+        registerBlock("incubator") { IncubatorBlock() }
+    val ADVANCED_INCUBATOR: DeferredBlock<Block> =
+        registerBlock("advanced_incubator") { AdvancedIncubatorBlock() }
 
     private fun <T : Block> registerBlock(
         name: String,