+++ CHANGELOG.md
@@ -23,3 +23,4 @@
 - The non-empty DNA Helices and Plasmids in the creative mode tab are now after everything else
 - Fixed Incubator and Advanced Incubator from not working (hasRecipe() returned false hasEnoughEnergy was true rather than when false)
 - The bubbles in the Incubator and Advanced Incubator now animate slower
+- Fixed broken tooltip for Basic Genes in the Plasmid Infuser